refactor(types): split api-constants.ts into domain files, export MS_PER_MINUTE
