+++ packages/types/src/api-constants/error-codes.ts
@@ -1,0 +1,54 @@
+/**
+ * API error code constants.
+ */
+
+// ── Error Codes ──────────────────────────────────────────────────────
+
+export const API_ERROR_CODES = {
+  VALIDATION_ERROR: "VALIDATION_ERROR",
+  INVALID_CURSOR: "INVALID_CURSOR",
+  INVALID_FRIEND_CODE: "INVALID_FRIEND_CODE",
+  UNAUTHENTICATED: "UNAUTHENTICATED",
+  SESSION_EXPIRED: "SESSION_EXPIRED",
+  KEY_VERSION_STALE: "KEY_VERSION_STALE",
+  FORBIDDEN: "FORBIDDEN",
+  SCOPE_INSUFFICIENT: "SCOPE_INSUFFICIENT",
+  BUCKET_ACCESS_DENIED: "BUCKET_ACCESS_DENIED",
+  NOT_FOUND: "NOT_FOUND",
+  CONFLICT: "CONFLICT",
+  INVALID_STATE: "INVALID_STATE",
+  IDEMPOTENCY_CONFLICT: "IDEMPOTENCY_CONFLICT",
+  HAS_DEPENDENTS: "HAS_DEPENDENTS",
+  ROTATION_IN_PROGRESS: "ROTATION_IN_PROGRESS",
+  FRIEND_CODE_EXPIRED: "FRIEND_CODE_EXPIRED",
+  CONNECTION_NOT_ACCEPTED: "CONNECTION_NOT_ACCEPTED",
+  BIOMETRIC_DISABLED: "BIOMETRIC_DISABLED",
+  INVALID_TOKEN: "INVALID_TOKEN",
+  INVALID_PIN: "INVALID_PIN",
+  PRECONDITION_FAILED: "PRECONDITION_FAILED",
+  BLOB_TOO_LARGE: "BLOB_TOO_LARGE",
+  QUOTA_EXCEEDED: "QUOTA_EXCEEDED",
+  RATE_LIMITED: "RATE_LIMITED",
+  LOGIN_THROTTLED: "LOGIN_THROTTLED",
+  ALREADY_ARCHIVED: "ALREADY_ARCHIVED",
+  NOT_ARCHIVED: "NOT_ARCHIVED",
+  ALREADY_PINNED: "ALREADY_PINNED",
+  NOT_PINNED: "NOT_PINNED",
+  ALREADY_ENDED: "ALREADY_ENDED",
+  ALREADY_RESPONDED: "ALREADY_RESPONDED",
+  ALREADY_DISMISSED: "ALREADY_DISMISSED",
+  POLL_CLOSED: "POLL_CLOSED",
+  TOO_MANY_VOTES: "TOO_MANY_VOTES",
+  ABSTAIN_NOT_ALLOWED: "ABSTAIN_NOT_ALLOWED",
+  VETO_NOT_ALLOWED: "VETO_NOT_ALLOWED",
+  SESSION_ARCHIVED: "SESSION_ARCHIVED",
+  INVALID_SUBJECT: "INVALID_SUBJECT",
+  INVALID_HIERARCHY: "INVALID_HIERARCHY",
+  CYCLE_DETECTED: "CYCLE_DETECTED",
+  MAX_DEPTH_EXCEEDED: "MAX_DEPTH_EXCEEDED",
+  UNSUPPORTED_MEDIA_TYPE: "UNSUPPORTED_MEDIA_TYPE",
+  INTERNAL_ERROR: "INTERNAL_ERROR",
+  SERVICE_UNAVAILABLE: "SERVICE_UNAVAILABLE",
+} as const;
+
+export type ApiErrorCode = (typeof API_ERROR_CODES)[keyof typeof API_ERROR_CODES];
